+++ contracts/interface/IScoreFunction.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.0;
 
 interface IScoreFunction {
     function calculateScore(uint256 distance, uint96 stake)
-        external returns (uint32);
+        external returns (uint40);
     function calculateDistance(uint8 hashFunctionId, bytes memory nodeId, bytes memory keyword)
         external returns (uint256);
 }
